requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4
dadaptation

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
neptune
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
autopep8        # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks
rich            # beautiful text formatting in terminal

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pytest          # tests
fiftyone        # fiftyone dataset
xxhash          # hash code for testings
# sh            # for running bash commands in some tests (linux/macos only)
